Organize each MongoDb node operations to its own files
